chore: rename cursor to predict
